% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{append_model}
\alias{append_model}
\title{append a model into apsimx}
\usage{
append_model(l, path, model)
}
\arguments{
\item{l}{the list of apsimx file}

\item{path}{If numeric, the path returned by search_path or search_node. If character, the path supported by apsimx}

\item{model}{A new model}
}
\value{
The modified list with new value
}
\description{
append a model into apsimx
}
