% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{available_models}
\alias{available_models}
\title{List all available models in APSIM NG}
\usage{
available_models()
}
\value{
a character vector of available models
}
\description{
List all available models in APSIM NG
}
\examples{
a <- available_models()
a[1:10]
}
