% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_apsim.R
\name{convert_xypair}
\alias{convert_xypair}
\title{Convert a model into xypair}
\usage{
convert_xypair(l)
}
\arguments{
\item{l}{the list of apsimx file}
}
\value{
a data.frame for X and Y values.
}
\description{
Convert a model into xypair
}
\examples{
m <- read_json('https://raw.githubusercontent.com/APSIMInitiative/ApsimX/master/Models/Resources/Wheat.json')
# Find the potential branching rate
potential <- search_path(model,
                         path = '[Structure].BranchingRate.PotentialBranchingRate.Vegetative.PotentialBranchingRate')
convert_xypair(potential)
}
