% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{disable_models}
\alias{disable_models}
\title{Disable models in apsimx}
\usage{
disable_models(l, paths)
}
\arguments{
\item{l}{the list of apsimx file}

\item{paths}{If numeric, the path returned by search_path or search_node. If character, the path supported by apsimx}
}
\value{
The modified list with new value
}
\description{
Disable models in apsimx
}
\examples{
wheat <- read_apsimx(system.file("extdata/Wheat.json", package = "rapsimng"))
a <- disable_models(wheat, '[Wheat].Phenology.ThermalTime')
}
