% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_element.R
\name{find_element}
\alias{find_element}
\title{Find an element in the apsimx file}
\usage{
find_element(l, ...)
}
\arguments{
\item{l}{the list of apsimx file}

\item{...}{Other names arguments for property}
}
\value{
A list matching all criterias
}
\description{
Find an element in the apsimx file
}
\examples{
m <- read_json('https://raw.githubusercontent.com/APSIMInitiative/ApsimX/master/Models/Resources/Wheat.json')
# Find the potential branching rate
potential <- find_element(model,
            Name = 'PotentialBranchingRate',
            XProperty = "[Structure].LeafTipsAppeared")
}
