% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_element.R
\name{find_element_xypair}
\alias{find_element_xypair}
\title{Find a xypair element in the apsimx file}
\usage{
find_element_xypair(l, ...)
}
\arguments{
\item{l}{the list of apsimx file}

\item{...}{Other names arguments for property}
}
\value{
a data.frame for X and Y values.
}
\description{
Find a xypair element in the apsimx file
}
\examples{
m <- read_json('https://raw.githubusercontent.com/APSIMInitiative/ApsimX/master/Models/Resources/Wheat.json')
# Find the potential branching rate
potential <- find_element_xypair(model,
            Name = 'PotentialBranchingRate',
            XProperty = "[Structure].LeafTipsAppeared")
}
