% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimng.R
\name{fixed_value}
\alias{fixed_value}
\title{The fixed value in the apsimx}
\usage{
fixed_value(l)
}
\arguments{
\item{l}{the list of apsimx file}
}
\value{
Value of this node
}
\description{
The fixed value in the apsimx
}
