% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cultivar.R
\name{get_cultivar}
\alias{get_cultivar}
\title{Get all cultivar parameters in a model}
\usage{
get_cultivar(l, alias = TRUE)
}
\arguments{
\item{l}{The list of apsimx file}

\item{alias}{Whether to export alias}
}
\value{
A data frame for all cultivar parameters
}
\description{
Get all cultivar parameters in a model
}
\examples{
wheat <- read_apsimx(system.file("extdata/Wheat.json", package = "rapsimng"))
get_cultivar(wheat)
get_cultivar(wheat, alias = FALSE)

}
