% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiment}
\alias{get_experiment}
\title{Get an experiment node by name}
\usage{
get_experiment(apsimx, experiment)
}
\arguments{
\item{apsimx}{A parsed APSIMX model from `read_apsimx()`.}

\item{experiment}{A single experiment name to locate.}
}
\value{
An `apsimxNode` object for the experiment.
}
\description{
Get an experiment node by name
}
\examples{
wheat <- read_apsimx(system.file("extdata/wheat.apsimx", package = "rapsimng"))
exp <- search_path(wheat, path = "[Experiment]")
get_experiment(wheat, exp$node$Name)
}
