% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{get_metfile}
\alias{get_metfile}
\title{Get the met file name for an experiment}
\usage{
get_metfile(l)
}
\arguments{
\item{l}{A list or apsimxNode red by read_apsimx}
}
\value{
The met file name in a experiment
}
\description{
Get the met file name for an experiment
}
\examples{
wheat <- read_apsimx(system.file("wheat.apsimx", package = "rapsimng"))
exp <- search_path(wheat, path = "[Experiment]")
get_metfile(exp)
}
