% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_tree.R
\name{get_parent}
\alias{get_parent}
\title{Get the parent node from a path}
\usage{
get_parent(l, path)
}
\arguments{
\item{l}{the list of apsimx file}

\item{path}{If numeric, the path returned by search_path or search_node. If character, the path supported by apsimx}
}
\value{
A new list for parent
}
\description{
Get the parent node from a path
}
\examples{
wheat <- read_apsimx(system.file("extdata/Wheat.json", package = "rapsimng"))
a <- search_path(wheat, '[Structure].BranchingRate')
get_parent(wheat, a$path)
}
