% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{insert_model}
\alias{insert_model}
\title{Insert a model into apsimx}
\usage{
insert_model(l, path, model)
}
\arguments{
\item{l}{the list of apsimx file}

\item{path}{If numeric, the path returned by search_path or search_node. If character, the path supported by apsimx}

\item{model}{A new model}
}
\value{
The modified list with new value
}
\description{
Insert a model into apsimx
}
