% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{list_report}
\alias{list_report}
\title{List all reports in the database}
\usage{
list_report(file)
}
\arguments{
\item{file}{the file path to apsimx or db file}
}
\value{
a vector of all reports
}
\description{
List all reports in the database
}
\examples{
\dontrun{
file <- system.file("extdata/wheat.apsimx", package = "rapsimng")
list_report(file)
}
}
