% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_level}
\alias{log_level}
\title{Set the log level of apsimx file}
\usage{
log_level(l, level = c("Error", "Warning", "Information", "Diagnostic", "All"))
}
\arguments{
\item{l}{the list of apsimx file}

\item{level}{log level with option Error, Warning, Information, Diagnostic, All}
}
\value{
a new apsimx file
}
\description{
Set the log level of apsimx file
}
