% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{minimum_apsimng}
\alias{minimum_apsimng}
\title{Create the minimum requirements to run an APSIM Next Generation}
\usage{
minimum_apsimng(install_path, output)
}
\arguments{
\item{install_path}{The installed path of APSIM Next Generation}

\item{output}{The output folder}
}
\description{
Create the minimum requirements to run an APSIM Next Generation
}
\examples{
\dontrun{
 minimum_apsimng("C:/ProgramFiles/APSIMNG", "minimum_apsimng")
}
}
