% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cultivar.R
\name{new_cultivar}
\alias{new_cultivar}
\title{Generate new cultivars with parameter which can be used in Replacements}
\usage{
new_cultivar(df, cultivar_folder = "Cultivars")
}
\arguments{
\item{df}{A data frame for new parameters with three columns, i.e. name, parameter and value.}

\item{cultivar_folder}{folder name for cultivars}
}
\value{
An APSIMX list
}
\description{
Generate new cultivars with parameter which can be used in Replacements
}
\examples{
df <- data.frame(name = c("C1", "C1", "C2", "C2"),
     parameter = c("[Phenology].CAMP.FLNparams.MinLN",
              "[Phenology].CAMP.FLNparams.VrnLN",
              "[Phenology].CAMP.FLNparams.MinLN",
              "[Phenology].CAMP.FLNparams.VrnLN"),
     value = c(5, 6, 7, 8))
a <- new_cultivar(df)
}
