% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{new_model}
\alias{new_model}
\title{Create a new model}
\usage{
new_model(name, namespace = "PMF")
}
\arguments{
\item{name}{The name of new model}

\item{namespace}{The namespace in the model}
}
\value{

}
\description{
Create a new model
}
\examples{
new_model("Cultivar")
}
