% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{read_report}
\alias{read_report}
\title{Read apsimx database in db file format}
\usage{
read_report(file, report)
}
\arguments{
\item{file}{the file path to apsimx or db file}

\item{report}{the report name}
}
\value{
a data.frame for a report
}
\description{
Read apsimx database in db file format
}
\examples{
file <- system.file("extdata/wheat.apsimx", package = "rapsimng")
read_report(file, "HarvestReport")
}
