% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{replace_model}
\alias{replace_model}
\title{Replace a model with new values}
\usage{
replace_model(l, path, model)
}
\arguments{
\item{l}{the list of apsimx file}

\item{path}{If numeric, the path returned by search_path or search_node. If character, the path supported by apsimx}

\item{model}{A new model}
}
\value{
The modified list with new value
}
\description{
Replace a model with new values
}
