% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_apsim.R
\name{search_path}
\alias{search_path}
\title{Find a model in the apsimx file using specified path}
\usage{
search_path(l, path)
}
\arguments{
\item{l}{the list of apsimx file}

\item{path}{The specified path (See details)}
}
\value{
The list for the specified path.
}
\description{
Find a model in the apsimx file using specified path
}
\section{Absolute Paths}{

Absolute paths have a leading ‘.’ e.g.
\itemize{
  \item .Simulations.Test.Clock - absolute path - refers to the clock model
in the ‘Test’ simulation.
}
}

\section{Scoped Paths}{

Scoped paths have a leading model type in square brackets. A model
of the specified name, in scope, is located before applying the rest of the path.
\itemize{
   \item [Soil].Water - scoped path - refers to the Water model that is a child of a model
that has the name ‘Soil’ that is in scope
}
}

\examples{
library(jsonlite)
m <- read_json('https://raw.githubusercontent.com/APSIMInitiative/ApsimX/master/Models/Resources/Wheat.json')
# Find the potential branching rate
potential <- search_path(m,
                         path = '[Structure].BranchingRate.PotentialBranchingRate.Vegetative.PotentialBranchingRate')
}
