% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_apsimx.R
\name{split_apsimx}
\alias{split_apsimx}
\title{Split an APSIMX file into separate simulations except cultivar}
\usage{
split_apsimx(file, output)
}
\arguments{
\item{file}{Character. Path to the input APSIMX file.}

\item{output}{Character. Path to the directory where the split components will be saved.}
}
\value{
No return value. The function is called for its side effects.
}
\description{
This function takes an APSIMX file and splits it into separate simulations for all factors except cultivar factors.
Then saving the results to the specified output location.
}
\examples{
\dontrun{
split_apsimx("path/to/input.apsimx", "path/to/output/directory")
}

}
