% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx_build.R
\name{test_apsimx}
\alias{test_apsimx}
\title{Test whether all files under published folder of apsimx are required}
\usage{
test_apsimx(base, example)
}
\arguments{
\item{base}{the base folder path to apsimx publish}

\item{example}{an example apsimx file}
}
\value{
A vector a required files
}
\description{
Test whether all files under published folder of apsimx are required
}
