% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_apsim.R
\name{write_apsimx}
\alias{write_apsimx}
\title{Write APSIMX file}
\usage{
write_apsimx(l, file)
}
\arguments{
\item{l}{the list of apsimx file}

\item{path}{The file path to apsimx file}
}
\value{
A list object of apsimx file
}
\description{
Write APSIMX file
}
