% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{xypair}
\alias{xypair}
\title{Convert a model into xypair}
\usage{
xypair(l, lower = NULL, upper = NULL)
}
\arguments{
\item{l}{the list of apsimx file}

\item{lower}{The lower bound. The first value will be copied lower is not NULL}

\item{upper}{The upper bound. The last value will be copied upper is not NULL}
}
\value{
a data.frame for X and Y values.
}
\description{
Convert a model into xypair
}
\examples{
file <- system.file("wheat.apsimx", package = "rapsimng")
m <- read_apsimx(file)
potential <- search_path(m,
   path = '[Structure].BranchingRate.PotentialBranchingRate.Vegetative.PotentialBranchingRate')
xypair(potential)
}
