% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{get_metfile}
\alias{get_metfile}
\title{Get the met file name for an experiment}
\usage{
get_metfile(l, is_stop = TRUE)
}
\arguments{
\item{l}{A list or apsimxNode red by read_apsimx}

\item{is_stop}{Whether stop the function when error}
}
\value{
The met file name in a experiment
}
\description{
Get the met file name for an experiment
}
\examples{
wheat <- read_apsimx(system.file("extdata/wheat.apsimx", package = "rapsimng"))
exp <- search_path(wheat, path = "[Experiment]")
get_metfile(exp)
}
