% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{new_model}
\alias{new_model}
\title{Create a new model}
\usage{
new_model(model, name = model)
}
\arguments{
\item{model}{The name of new model}

\item{name}{The new name}
}
\description{
Create a new model
}
\examples{
new_model(model = "PMF.Cultivar")
new_model(model = "PMF.Cultivar", name = "example")
}
