% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_apsim.R
\name{read_apsimx}
\alias{read_apsimx}
\title{Read APSIMX file}
\usage{
read_apsimx(path)
}
\arguments{
\item{path}{The file path or URL to apsimx file}
}
\value{
A list object of apsimx file
}
\description{
Read APSIMX file
}
\examples{
m <- read_apsimx('https://raw.githubusercontent.com/APSIMInitiative/ApsimX/master/Models/Resources/Wheat.json')
}
