% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_apsimx}
\alias{read_apsimx}
\title{Read APSIMX file}
\usage{
read_apsimx(path)
}
\arguments{
\item{path}{The file path or URL to apsimx file}
}
\value{
A list object of apsimx file
}
\description{
Read APSIMX file
}
\examples{
file <- system.file("extdata/wheat.apsimx", package = "rapsimng")
m <- read_apsimx(file)
}
