% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{remove_model}
\alias{remove_model}
\title{Remove a model with new values}
\usage{
remove_model(l, path)
}
\arguments{
\item{l}{the list of apsimx file}

\item{path}{If numeric, the path returned by search_path or search_node. If character, the path supported by apsimx}
}
\value{
The modified list with new value
}
\description{
Remove a model with new values
}
\examples{
wheat <- read_apsimx(system.file("extdata/Wheat.json", package = "rapsimng"))
a <- search_path(wheat, '[Wheat].Phenology.ThermalTime')
wheat_new <- remove_model(wheat, a$path)
b <- search_path(wheat_new, '[Wheat].Phenology.ThermalTime')
b
}
