% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_apsim.R
\name{replace_model}
\alias{replace_model}
\title{Replace a model with new values}
\usage{
replace_model(l, path, new)
}
\arguments{
\item{l}{the list of apsimx file}

\item{path}{The path returned by search_path or search_node}

\item{new}{A new model}
}
\value{
The modified list with new value
}
\description{
Replace a model with new values
}
