% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Run apsimx file using Models.exe}
\usage{
run_models(
  models_exe,
  path,
  pattern = NULL,
  recurse = FALSE,
  csv = FALSE,
  parallel = NULL,
  ncpus = NULL,
  verbose = FALSE
)
}
\arguments{
\item{models_exe}{path to Models.exe}

\item{path}{The path to an .apsimx file. May include wildcard.}

\item{pattern}{Use to filter simulation names to run.}

\item{recurse}{Recursively search subdirectories for files matching ApsimXFileSpec. FALSE in default.}

\item{csv}{Export all reports to .csv files. FALSE in default.}

\item{parallel}{Use the multi-process job runner. If FALSE, use single threaded; if TRUE, use the multi-process job runner}

\item{ncpus}{Set the number of processors to use. All processes in default}

\item{verbose}{Write messages to StdOut when a simulation starts/finishes. Only has an effect when running a directory of .apsimx files (*.apsimx).}
}
\description{
Run apsimx file using Models.exe
}
