% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimng.R
\name{search_node}
\alias{search_node}
\title{Find an element in the apsimx file (first occurrence)}
\usage{
search_node(l, max_depth = 1e+06, ...)
}
\arguments{
\item{l}{the list of apsimx file}

\item{max_depth}{The maximum depth to search}

\item{...}{Other names arguments for property}
}
\value{
A list matching all criterias
}
\description{
Find an element in the apsimx file (first occurrence)
}
\examples{
library(jsonlite)
m <- read_json('https://raw.githubusercontent.com/APSIMInitiative/ApsimX/master/Models/Resources/Wheat.json')
# Find the potential branching rate
potential <- search_node(m,
            Name = 'PotentialBranchingRate',
            XProperty = "[Structure].LeafTipsAppeared")
}
