% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_apsim.R
\name{search_node}
\alias{search_node}
\title{Find an element in the apsimx file}
\usage{
search_node(l, ...)
}
\arguments{
\item{l}{the list of apsimx file}

\item{...}{Other names arguments for property}
}
\value{
A list matching all criterias
}
\description{
Find an element in the apsimx file
}
\examples{
m <- read_json('https://raw.githubusercontent.com/APSIMInitiative/ApsimX/master/Models/Resources/Wheat.json')
# Find the potential branching rate
potential <- search_node(model,
            Name = 'PotentialBranchingRate',
            XProperty = "[Structure].LeafTipsAppeared")
}
